Refactor prng to not use 64-bit atomics on 32-bit platforms.
This resolves #495.

+++ include/jemalloc/internal/arena.h
@@ -182,7 +182,7 @@ struct arena_s {
 	 * PRNG state for cache index randomization of large allocation base
 	 * pointers.
 	 */
-	uint64_t		offset_state;
+	size_t			offset_state;
 
 	dss_prec_t		dss_prec;
